Flytt identisk valideringsfunksjon for begrunnelse til felles utilfil
