+++ src/frontend/komponenter/Fagsak/Vilkårsvurdering/GeneriskVilkår/VilkårSkjema.tsx
@@ -269,9 +269,11 @@ export const erBegrunnelsePåkrevd = (
     vurderesEtter: Regelverk | undefined,
     utdypendeVilkårsvurderinger: UtdypendeVilkårsvurdering[],
     personType: PersonType,
-    vilkårType: VilkårType
+    vilkårType: VilkårType,
+    søkerHarMeldtFraOmBarnehageplass?: boolean
 ): boolean => {
     return (
+        (vilkårType == VilkårType.BARNEHAGEPLASS && søkerHarMeldtFraOmBarnehageplass) ||
         (vurderesEtter === Regelverk.NASJONALE_REGLER && utdypendeVilkårsvurderinger.length > 0) ||
         (vurderesEtter === Regelverk.EØS_FORORDNINGEN &&
             personType === PersonType.SØKER &&
